seprate imageuploader component
